Struct safety_rules::process_client_wrapper::ProcessClientWrapper

pub struct ProcessClientWrapper<T> { /* fields omitted */ }

This container exists only so that we can kill the spawned process after testing is complete. Otherwise the process will be killed at the end of the safety_rules function and the test will fail.

Methods

impl<T: Payload> ProcessClientWrapper<T>

pub fn new(backend: SecureBackend) -> Self

pub fn signer(&self) -> ValidatorSigner