Struct safety_rules::PersistentSafetyStorage

pub struct PersistentSafetyStorage { /* fields omitted */ }

SafetyRules needs an abstract storage interface to act as a common utility for storing persistent data to local disk, cloud, secrets managers, or even memory (for tests) Any set function is expected to sync to the remote system before returning. @TODO add access to private key from persistent store @TODO add retrieval of private key based upon public key to persistent store

Methods

impl PersistentSafetyStorage

pub fn in_memory(private_key: Ed25519PrivateKey) -> Self

pub fn initialize(
internal_store: Box<dyn Storage>,
private_key: Ed25519PrivateKey,
waypoint: Waypoint
) -> Self

Use this to instantiate a PersistentStorage for a new data store, one that has no SafetyRules values set.

pub fn new(internal_store: Box<dyn Storage>) -> Self

Use this to instantiate a PersistentStorage with an existing data store. This is intended for constructed environments.

pub fn consensus_key(&self) -> Result<Ed25519PrivateKey>

pub fn set_consensus_key(
&mut self,
consensus_key: Ed25519PrivateKey
) -> Result<()>

pub fn epoch(&self) -> Result<u64>

pub fn set_epoch(&mut self, epoch: u64) -> Result<()>

pub fn last_voted_round(&self) -> Result<Round>

pub fn set_last_voted_round(&mut self, last_voted_round: Round) -> Result<()>

pub fn preferred_round(&self) -> Result<Round>

pub fn set_preferred_round(&mut self, preferred_round: Round) -> Result<()>

pub fn waypoint(&self) -> Result<Waypoint>

pub fn set_waypoint(&mut self, waypoint: &Waypoint) -> Result<()>