Struct safety_rules::SafetyRules

pub struct SafetyRules { /* fields omitted */ }

SafetyRules is responsible for two things that are critical for the safety of the consensus:

voting rules,

commit rules. SafetyRules is NOT THREAD SAFE (should be protected outside via e.g., RwLock). The commit decisions are returned to the caller as result of learning about a new QuorumCert.

Methods

impl SafetyRules

pub fn new(
state: ConsensusState,
validator_signer: Arc<ValidatorSigner>
) -> Self

Constructs a new instance of SafetyRules given the BlockTree and ConsensusState.

pub fn signer(&self) -> &ValidatorSigner

pub fn update(&mut self, qc: &QuorumCert)

Learn about a new quorum certificate. Several things can happen as a result of that:

update the preferred block to a higher value.

commit some blocks. In case of commits the last committed block is returned. Requires that all the ancestors of the block are available for at least up to the last committed block, might panic otherwise. The update function is invoked whenever a system learns about a potentially high QC.

pub fn consensus_state(&self) -> ConsensusState

Clones the up-to-date state of consensus (for monitoring / debugging purposes)

pub fn construct_and_sign_vote<T: Payload>(
&mut self,
vote_proposal: &VoteProposal<T>
) -> Result<Vote, Error>

Attempts to vote for a given proposal following the voting rules. The returned value is then going to be used for either sending the vote or doing nothing. In case of a vote a cloned consensus state is returned (to be persisted before the vote is sent). Requires that all the ancestors of the block are available for at least up to the last committed block, might panic otherwise.

pub fn sign_proposal<T: Payload>(
&self,
block_data: BlockData<T>
) -> Result<Block<T>, Error>

pub fn sign_timeout(&self, timeout: &Timeout) -> Result<Signature, Error>

As the holder of the private key, SafetyRules also signs what is effectively a timeout message. This returns the signature for that timeout message.