safety_rules

Trait TSafetyRules

pub trait TSafetyRules {
    // Required methods
    fn consensus_state(&mut self) -> Result<ConsensusState, Error>;
    fn initialize(&mut self, epoch_state: &EpochState) -> Result<(), Error>;
    fn construct_and_sign_vote(
        &mut self,
        vote_proposal: &MaybeSignedVoteProposal,
    ) -> Result<Vote, Error>;
    fn sign_proposal(&mut self, block_data: BlockData) -> Result<Block, Error>;
    fn sign_timeout(
        &mut self,
        timeout: &Timeout,
    ) -> Result<ConsensusSignature, Error>;

    // Provided methods
    fn start_voting(&mut self, _initialize: bool) -> Result<(), Error> { ... }
    fn stop_voting(&mut self) -> Result<(), Error> { ... }
}

Interface for SafetyRules

Required Methods

fn consensus_state(&mut self) -> Result<ConsensusState, Error>

Provides the internal state of SafetyRules for monitoring / debugging purposes. This does not include sensitive data like private keys.

fn initialize(&mut self, epoch_state: &EpochState) -> Result<(), Error>

Initialize SafetyRules with the given epoch state. Sets up the validator set and signing keys for the epoch.

fn construct_and_sign_vote( &mut self, vote_proposal: &MaybeSignedVoteProposal, ) -> Result<Vote, Error>

Attempts to vote for a given proposal following the voting rules.

fn sign_proposal(&mut self, block_data: BlockData) -> Result<Block, Error>

As the holder of the private key, SafetyRules also signs proposals or blocks. A Block is a signed BlockData along with some additional metadata.

fn sign_timeout( &mut self, timeout: &Timeout, ) -> Result<ConsensusSignature, Error>

As the holder of the private key, SafetyRules also signs what is effectively a timeout message. This returns the signature for that timeout message.

Provided Methods

fn start_voting(&mut self, _initialize: bool) -> Result<(), Error>

Allow the safety rule to start voting with saved secure data from another node.

fn stop_voting(&mut self) -> Result<(), Error>

Stop the safety rule from voting and save secure data.