Crate safety_rules

Modules

process_client_wrapper

test_utils

Structs

ConsensusState
Public representation of the internal state of SafetyRules for monitoring / debugging purposes. This does not include sensitive data like private keys. @TODO add hash of ledger info (waypoint)

PersistentSafetyStorage
SafetyRules needs an abstract storage interface to act as a common utility for storing persistent data to local disk, cloud, secrets managers, or even memory (for tests) Any set function is expected to sync to the remote system before returning. @TODO add access to private key from persistent store @TODO add retrieval of private key based upon public key to persistent store

Process

SafetyRules
SafetyRules is responsible for the safety of the consensus:

SafetyRulesManager

Enums

Error
Different reasons for proposal rejection

Statics

COUNTERS

Traits

TSafetyRules
Interface for SafetyRules