Crate safety_rules

Re-exports

pub use crate::fuzzing_utils::fuzzing;

Modules

fuzzing_utils

test_utils

Structs

ConsensusState

Public representation of the internal state of SafetyRules for monitoring / debugging purposes. This does not include sensitive data like private keys. @TODO add hash of ledger info (waypoint)

PersistentSafetyStorage

SafetyRules needs an abstract storage interface to act as a common utility for storing persistent data to local disk, cloud, secrets managers, or even memory (for tests) Any set function is expected to sync to the remote system before returning.

Process

SafetyRules

@TODO consider a cache of verified QCs to cut down on verification costs

SafetyRulesManager

Enums

Error

Different reasons for proposal rejection

Traits

TSafetyRules

Interface for SafetyRules