Struct safety_rules::SafetyRules

pub struct SafetyRules<T> { /* fields omitted */ }

SafetyRules is responsible for the safety of the consensus:

voting rules

commit rules

ownership of the consensus private key @TODO add a benchmark to evaluate SafetyRules @TODO consider a cache of verified QCs to cut down on verification costs @TODO bootstrap with a hash of a ledger info (waypoint) that includes a validator set @TODO update storage with hash of ledger info (waypoint) during epoch changes (includes a new validator set)

Methods

impl<T: Payload> SafetyRules<T>

pub fn new(author: Author, persistent_storage: PersistentSafetyStorage) -> Self

Constructs a new instance of SafetyRules with the given persistent storage and the consensus private keys @TODO replace this with an API that takes in a SafetyRulesConfig

pub fn construct_ledger_info(&self, proposed_block: &Block<T>) -> LedgerInfo

Produces a LedgerInfo that either commits a block based upon the 3-chain commit rule or an empty LedgerInfo for no commit. The 3-chain commit rule is: B0 (as well as its prefix) can be committed if there exist certified blocks B1 and B2 that satisfy:

B0 <- B1 <- B2 <--

round(B0) + 1 = round(B1), and

round(B1) + 1 = round(B2).